pyproject.toml
[tool.poetry]
name = "vendaflex"
version = "0.1.0"
description = ""
authors = ["Christian Catrofe <christianlopes100@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.109.0"
uvicorn = "^0.25.0"
asyncpg = "^0.29.0"
pydantic-settings = "^2.1.0"
bcrypt = "^4.1.2"
sqlalchemy = "^2.0.25"
pyjwt = "^2.8.0"


[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"
black = "^23.12.1"
flake8 = "^7.0.0"
mypy = "^1.8.0"
pytest = "^7.4.4"
pre-commit = "^3.6.0"
pytest-random-order = "^1.1.0"
pytest-cov = "^4.1.0"
aiosqlite = "^0.19.0"
httpx = "^0.26.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
